requirements.txt
beautifulsoup4==4.12.3
click==8.1.7
google==3.0.0
numpy==2.0.1
protobuf==4.25.4
protoletariat==3.2.19
pyserial==3.5
setuptools==72.2.0
soupsieve==2.6
keyboard == 0.13.5
